Simplifying coprocessing plugin wizard.
We are trying to encourage the convention that the input to the
coprocessing library is called "input" to identity the source.
For the coprocessing plugin wizard used to generate the script
the user had to specify which input was called "input" but for
only a single input/source this just made the user have to type
extra stuff that was annoying.  Now if there is only a single
source/input it is automatically given the "input" name
for the wizard.

+++ Plugins/CoProcessingScriptGenerator/pqCPExportStateWizard.cxx
@@ -245,7 +245,16 @@ void pqCPExportStateWizardPage3::initializePage()
     QListWidgetItem* item = this->Internals->simulationInputs->item(cc);
     QString text = item->text();
     this->Internals->nameWidget->setItem(cc, 0, new QTableWidgetItem(text));
-    this->Internals->nameWidget->setItem(cc, 1, new QTableWidgetItem(text));
+    // if there is only 1 input then call it input, otherwise
+    // use the same name as the filter
+    if(this->Internals->simulationInputs->count() == 1)
+      {
+      this->Internals->nameWidget->setItem(cc, 1, new QTableWidgetItem("input"));
+      }
+    else
+      {
+      this->Internals->nameWidget->setItem(cc, 1, new QTableWidgetItem(text));
+      }
     QTableWidgetItem* tableItem = this->Internals->nameWidget->item(cc, 1);
     tableItem->setFlags(tableItem->flags()|Qt::ItemIsEditable);
